[level 0] Title: 등수 매기기, Time: 0.02 ms, Memory: 74.4 MB -BaekjoonHub

+++ 프로그래머스/lv0/120882. 등수 매기기/등수 매기기.java
@@ -5,7 +5,7 @@ class Solution {
         
         // 각 학생의 평균 점수를 저장(double형)
         for(int i = 0; i < score.length; i++)
-            avg[i] = (double) (score[i][0] + score[i][1]) / 2.0;
+            avg[i] = (score[i][0] + score[i][1]) / 2.0;
         
         // 각 학생의 평균 점수를 비교, 등수를 저장
         for(int i = 0; i < avg.length; i++) {
